+++ docs/changelog.md
@@ -27,6 +27,10 @@
 
 ## 更新说明
 
+1.5.90 fix: fix macro_china_bond_public interface
+
+    1. 修复 macro_china_bond_public 接口
+
 1.5.89 fix: fix stock_info_sh_name_code interface
 
     1. 修复 stock_info_sh_name_code 接口
@@ -639,6 +643,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.5.90 fix: fix macro_china_bond_public interface
 
 1.5.89 fix: fix stock_info_sh_name_code interface